Separated growing of the blit and index buffers.  Adjusted the data types of the buffer sizes to match their indices and added a cap for the max sizes.

+++ trunk/SDL_gpu/OpenGL_1_BASE/SDL_gpu_OpenGL_1_BASE.h
@@ -49,11 +49,11 @@ typedef struct ContextData_OpenGL_1_BASE
 	GPU_Image* last_image;
 	GPU_Target* last_target;
 	float* blit_buffer;  // Holds sets of 4 vertices and 4 tex coords interleaved (e.g. [x0, y0, z0, s0, t0, ...]).
-	unsigned int blit_buffer_num_vertices;
-	unsigned int blit_buffer_max_num_vertices;
+	unsigned short blit_buffer_num_vertices;
+	unsigned short blit_buffer_max_num_vertices;
 	unsigned short* index_buffer;  // Indexes into the blit buffer so we can use 4 vertices for every 2 triangles (1 quad)
-	int index_buffer_num_vertices;
-	int index_buffer_max_num_vertices;
+	unsigned int index_buffer_num_vertices;
+	unsigned int index_buffer_max_num_vertices;
 } ContextData_OpenGL_1_BASE;
 
 typedef struct RendererData_OpenGL_1_BASE
